requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
openai==1.57.2
PyMuPDF==1.24.14
faiss-cpu==1.9.0.post1
numpy==1.26.4
python-dotenv==1.0.1
